requirements.txt
Flask==2.0.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
blinker==1.4
click==8.0.3
gunicorn==20.1.0
itsdangerous==2.0.1
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2-binary==2.9.3
Werkzeug==2.0.2
python-dotenv==0.19.2
pep517==0.12.0
pip-tools==6.4.0
PyJWT==2.3.0
tomli==2.0.0
